requirements.txt
# Core - Meraki SDK
meraki>=1.53.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# CLI & Output
click>=8.1.0
rich>=13.0.0

# Templates & Reports
jinja2>=3.1.0
weasyprint>=60.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
